Fix: don't complete when in comment or string
Currently when `company-hy` is asked for a `prefix` it just calls `(company-grab-symbol)`.
If this happens inside the string no meaningful completion can be done and `company-hy` should return `nil`.
This PR adds a check for this with `(company-in-string-or-comment)`, thus giving option to other backbends to do their work.

Most notable this allows file completion via `company-files` inside strings, so
```hy
(pd.read-csv "./") ;; will complete files instead of doing nothing
```